requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.14.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.2
protobuf==4.25.3
pyarrow==15.0.2
pydeck==0.8.1b0
Pygments==2.17.2
pylightning==0.0.7.3
python-bitcoinlib==0.12.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.32.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
watchdog==4.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
python-dotenv==0.19.1
